50 test fonctionnel + Uniformité du code + Correction de plein de fonctions

+++ CommuMoney/CommuMoney.METIER/Metier/Depenses_METIER.cs
@@ -13,10 +13,10 @@ namespace CommuMoney.METIER.Metier
         public int ID { get; set; }
         public int ID_Personne { get; set; }
         public int ID_Projet { get; set; }
-        public float Montant { get; set; }
+        public double Montant { get; set; }
 
-        public Depenses_METIER(int id_personne, int id_projet, float montant) => (ID_Personne, ID_Projet, Montant) = (id_personne, id_projet, montant);
-        public Depenses_METIER(int id, int id_personne, int id_projet, float montant) => (ID, ID_Personne, ID_Projet, Montant) = (id, id_personne, id_projet, montant);
+        public Depenses_METIER(int id_personne, int id_projet, double montant) => (ID_Personne, ID_Projet, Montant) = (id_personne, id_projet, montant);
+        public Depenses_METIER(int id, int id_personne, int id_projet, double montant) => (ID, ID_Personne, ID_Projet, Montant) = (id, id_personne, id_projet, montant);
 
         #region Insert
         public void Insert()
